Added WebsiteUI.fxml and icons for it
Created WebsiteUI.fxml which will display the user's credentials for the website. The top left of the the pane will contain the website name and the top right of the pane is an ImageView that will display the website's logo.

Added copyIcon.png for use in WebsiteUI.fxml for the copy buttons.

Added locksmithLogo.png for use in LoginPage.fxml for the login page.